backend/requirements.txt
django==5.2.5
djangorestframework==3.15.0
mongoengine==0.27.0
pymongo==4.6.1
python-decouple==3.8
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
PyJWT==2.8.0
